refactor: database on_error_rollback decorator moved to database driver
